requirements.txt
peft
torch
transformers==4.28.1
pytorch_lightning==1.9.0
Pillow
numpy
gradio
